Softlimits when jogging and DIR=1 did not work
When the DIR field is set to 1, jogging forward in EPICS
makes the real motor "turn backwards".
pmr->cdir is 0 and jvel send to the motor is negative.

The soft limit check did not work:
To find out into which direction the motor is moving and whether to check
high or low limit we need to look at pmr->cdir instead of pmr->jogf.