Add TCPSimulator, model3 driver, simulator and testscripts
Add 3 directories:
epicsIOC/     which includes a model 3 driver
simulator/    which has the simulator code
testScripts/  with a basic test script

The simulator is inspired by Kevin Peterson's VMC, and uses the same
syntax.
The ambition of this implementation is to extend it more and more
to be able to run all kinds of test scripts on a simulated motor.