When RA_PROBLEM is true, stop motor only if needed.
When the motor reports a problem, it should stop and
the problem needs to be analyzed.
The analyzes may need vendor specifig tools, which means
that the problem must be retained until the analyze is
completed, and the vendor specific tools may be needed
to put the motor back to normal condition.
Commit 303a9208e3 changed this to
'Set "stop" field true if driver returns RA_PROBLEM true.'

This means tha the motorRecord will send a stop command after each
poll, but the problem does not got away, and logs are flooded
with usesless transactions.
The stop feature may be needed for some controllers, so send
a stop when the controller does not return DONE = true.
(In general, this may be done in the driver for this very
controller)