VBAS unsupported
Whenever a controller does not use VBAS, but the field is
set to some value != 0.0 in the record, the calculation of the
acceleration parameters is wrong:

The record assumes that the axis jumps to VBAS in zero time, and ramps
up to e.g. VELO with the acceleration specified.

One improvement has already been mad when VELO == VBAS,
see commit b201e40ee3, "Changed ACCL calculation when VELO == VBAS."

Now we make another improvement: the driver can tell the record to set VBAS
to 0.0.

This is done by setting by setting bit 15 (counted from 0) in MSTA:
"This axis doesn't support VBAS at all, so please set it to 0".