autopower on "mode 2"
Inpmlement autopower mode 2:
When a movement is started, the power needs to be turned on.
- power on
- sleep
- start to move

Mode 2 allows to shorten the sleep(): If the drive can return
"power is on, go ahead", then the sleep can be aborted.

Note that drivers that can't read back the power on status reliable
simply return false in pollPowerIsOn().

If a drive does not allow to return the power on status at all,
the driver doesn't have to implement/overload pollPowerIsOn() since
the default implementation in asynMotorAxis always returns false.

autopower mode 2 allows to keep the power on forever, if the
PwrOffDelay is < 0.0