Use FRAC for relative movements
Re-implement the FRAC field, but probably in a different way than before.
Now the FRAC field is only used when the record commands relative movements.
And relative movements are used when there is an external feedback,
either an encoder or an EPICS PV (and RMOD != motorRMOD_I).

The question is, if this is the right thing to do.
The other question is, how BDST should work together with FRAC.

Update motor_release.html