Reset MISS when the movement may be started
When ever the VAL (or DVAL, RVAL) changes and the record pick up the
change to start a new movement, reset MISS.
This puts the record of an alarm state, if MISV is != 0.

Note:
Even when the motor does not need to be started because we are inside
the SDBD range, MISS is now reset.