motorRecord.cc: Home/Eternal move & DLY did give MISS
When the motor is homed, it may be at a certain position
after the homing.
When DLY and RTRY are set, this position is not the same as DVAL.
Make sure that the DLY is executed, but no retry is tried.

After the processing is done, sync VAL/DVAL with RBV/DRBV