asynMotorPoller: Handle asynStatus; clean up lock()
Clean up the lock()/unlock() handling, only take the lock
for a short time.
TODO:
Look at the poll() function, is anybody using it ?

At the same time, aboort the pollall() in case of asynError,
(that has been there before)
and set asynStatusConnected_ (this is new).
As a consequence, the poller  now wait for the connection to be connected,
before calling initialPoll().