enforceMinRetryDeadband(): Take values from controller
Whenever the controller has a value for RDBD/SDBD, put it into the record,
regardless if the record has a value != 0.0 or not.
This does not allow the overwriting in the record any more.
(Which does not make too much sense, since the controller needs to be
 configured with the right values to do postitioning).

On the advantage side, now a controller can be re-comissioned and restarted.
Now the record picks up the new values without the need for an IOC restart.