motorRecord.cc: RE_INIT must call enforceMinRetryDeadband()
Whenever the re-init is done, the SDBD and RDBD fields must be updated
with values from controller (if they are set of course).
This makes sure that changing the values in the controller followed by
a restart of the controller (which looses the connection and re-connects)
will update the record to the new values.

Add a call to enforceMinRetryDeadband() if process_reason == CALLBACK_RE_INIT.