motorRecord.cc: Jogging against or from LVIO
When jogging has stopped because the motor is close to a softlimit,
there was a problems
Pressing the same jog button again trying to move again into the
soft limit hanged the record.
STOP didn't help to unlock it, neither did jog into the opposite
direction work.

Solution:
Clear the JOG_REQ in the MIP and the button in the record.