motorRecord.cc: DLY and JOG gave MISS
When the DLY field was set to a value > 0.0, the state machine did not
handle the end of e.g. a JOG very well:
maybeRetry() was called and the MISS field was set to 1.
Change the DLY handling, once the delay is completed, reset both the delay
bits in MIP and make "pmr->mip != MIP_DONE" work as expected.

Note:
The delay handling may be refactured even more, but that will be done
in a different commit.