'E: Axis not homed' only when PROBLEM bit is set
The "E:" prefix should only be used, when the record
is put into alarm state. This is the case when the
problem bit is set.
Use "E: Axis not homed" when the problem bit is set,
but "W: Axis not homed" otherwise.
And yes, as usual, this feature must be enabled with help of
"motorNotHomedProblem_"