drvMotorAsyn.c: Remove enum warnings
There are 2 different enums, motorCommand and motorAxisParam_t,
which are mixed with each other.

The compiler doesn't like this mixture, and spits out a warning
when e.g. (*pPvt->drvset->getInteger) is called.
An enum "motorCommand" is expected, but "motorAxisParam_t" is provided.

Use "int" in some declarations instead, which does not give a warning,
or cast to "int" otherwise.