motorRecord.cc: init_record() did not handle alarm_sub()
When the record is initialized, init_record() is called.
Even when all data is known from the controller (RBV, MSTA, etc)
the alarms are not (yet) set.
When the motor is in a "PROBLEM" state, a process() is needed
to get the EPICS record alarms set.
Trigger a process() by calling GET_INFO via devSupGetInfo()