MsgTxt: Replace 'W:' with 'I:'
Don't show messages with the 'W:' prefix for warning any more.
Errors still have an 'E:' and should put the record into
the major alarm state.
Everything else is an information to the user, even if e.g.
linit violation may put the record into minor alarm.
This decision is done by the record and not by the driver, so
use the less frightening 'I:'