readbackEGU
Make it possible for a driver to report the readback values in EGU,
including the read-only-soft-limits.

There is a new parameter in the parameter library,
motorFlagsReadbackEGU_, named  "MOTOR_FLAGS_READBACK_EGU".
This bit is even found in the mflg field of the record.

When it is set, the following will happen:
- devMotorAsyn.c will calculate the RMP field (in motor steps)
  from the commanded motor position (in EGU).
- devMotorAsyn.c will calculate the REP field (in encoder ticks)
  from the encoder readback (in EGU).
- motorRecord.cc will calculate DRBV depending on UEIP,
  readBack.readbackEGU, and, if needed, MRES or ERES.

The good thing is that DRBV can now be retrieved directly from the
controller without any forth-and-back multiplications/divisions.
The raw values like RRBV are still supported.
if a legacy application needs them.