motorRecord.cc: Soft limits from controller 'as is'
When the "read only" soft limits in the controller are updated,
push them into DHLM/DLLM (and HLM/LLM).
The old code did that only when the new values had been "tighter"
than what we have in the record.
Doing so gives problems when the axis is commissioned only,
and soft limits are adjusted to both directions, but the
changes are not pushed into the record.