Don't allow DHLM=DLLM=0 when read-only limits are set
Whenever the controller defines read only softlimits,
the values for DHLM and DLLM should stay inside this range.

Therefore we check that DHLM < maxValue and DLLM > minValue
before applying the values in the record.

There are however, bordercases.
With minValue = 0.0 and maxValue = 165, it is allowed
to set DHLM to 0.0 and DLLM to 0.0
This makes the record think that there are now limits,
and not allowed values may be commanded to the controller.

Avoid this by not allowing DHLM == DLLM == 0 whenever
softLimitRO.motorDialLimitsValid is set.