motorRecord.cc: Correct SPDB handling when MRES == 1.0
For a soft motor (or even a real motor) MRES may be 1.0
and SPDB 0.1. The old logic did look at MRES first, to
decide if a move should be done.
Tweaking a motor with 1 mm may not always work, some tweaks
are not executed when MRES == 1 and the motor is at 0.990
and should be tweaked to 2.000

Solution:
If SPDB is > 0 ignore MRES.