motorRecord.cc: Print the date and time in Debug()
Major improvements in the Debug() function.
Convert the DEbug() function into a macro, so that the
__FILE__ and __LINE__ don't have to be added as a parameter
any more when using it.
The functionality is moved into the function mrPrint(),
which does the following nice things:

Get the date and time, and add it to the print caused by Debug().
This is similar to what asyn does in the prints.