Improve logging of DHLM/DLLM changes (softLimitRO)
When DHLM/DLLM is set (or HLM/LLM), the read only soft limits from the
controller are respected and limit the allowed range.
E.g. the controller has a read only soft limit of 165.0 for DHLM, then
a value of 170.0 will be clipped to 165.0
Make more clear where this clipping is coming from,
and call it "softLimitRO" in the IOC log.