Late connect: Don't mis-use UDF, use neverPolled.
If the controller is "offline" when the record is initialized,
the "readback value" can not be put into .VAL field.
To be more exact, the DVAL field can not be initialized from DRBV,
since .DRBV is undefined.

The "never polled, record init never completed" feature needs a variable,
which in is lifetime is changed only once.
So far, the UDF field has been used.
However, UDF is connectect to a (possible) input link, and may change
from TRUE to FALSE and vice versa.
De-couple UDF from  "late connect" and use a private variable
"neverPolled" instead.