motorRecord.cc: Improve printing of MIP
Because the MIP has so many bits that may change at the same time, they all are
abbreviated to 2 letters to make printing short-ish.
However, if only one bit is changed, we have place to print a longer string.
Example: The "JOG_STOP" was abbreviated to "jS".
Now, if only "jS" is set, we print "JOG_STOP(jS)" to make clear, what "jS" stands for.
This make it still possible to search for "jS".