motorRecord.cc: Reset JOGF/JOGR in motor-has-stopped
Reset the JOGF and JOGR when the motor has stopped, being busy
with "JOG_STOP" or one of the "JOG_Backlash" MIP states.

This prevents the record to stay in a state where one of JOGF or JOGR
is set, but the motor stands still and has no intention to jog at all.

This change resets both JOGF and JOGR (instead of ignoring them).

It covers 2 scenarios:
- While JOGF is active, JOGR is activated as well.
  Solution: Stop the motor, reset both fields, and let the user
  press one of those buttons.
- While still doing backlash caused by jogging, the user activates
  JOGF or JOGR again.
  Ignore this request, the user can issue a new one once the motor is DONE.