motorRecord.cc: clear_buttons() when hitting a limit switch
The logic in the motorRecord says that another "update" is needed when a motor
reports "done" and is stitting on a limit switch (in the commanded direction).

Under this circumstances it can happen, that HOMF, JOGF, HOMR or JOGR stay active
in the record.
Fix this by clearing all buttons.