motorRecord.cc: Less LS special handling with cdir
The original motors used with motors did only have one one bit
to signal "limit switch active". Both high- and low had been
mapped to a single bit.
As a result, the ".CDIR" field had been introduced to distinguish
between the limit switches and have some logigc here.
Looking back, this logic should be in the motion controller firmware.
Or, if that is not possible, be hidden inside the motor-specific driver.
Or, if more than one or two drivers need it, be in the common code,
and made configurable.

This is more important with "modern" comtrollers, where e.g. HOMF and HOMR
do the very same thing.
They home into a controller-defined way. In that sense the .CDIR field
may be completely wrong inside the motorRecord.

Beside that, when modern controllers return "DONE", they are done.

In short, #ifdef away the code that may confuse our state-machine around
the .MIP field. Leave the code visible, in case we want to re-active it.