motorRecord.cc: Introduce mr_stops_on_ls_activated()
The original motorRecord stops the motor, when a limit switch is hit.
There is an extra filter, the .CDIR field, to stop only on the limit
switch for the "commanded direction".
Modern controllers do the stop on their own, and may be confused by an
additional stop.
The driver can cofigure the motorRecord stop away: Activate MF_NOSTOP_ONLS

Another case when we don't want the motorRecord to stop the motor is when
we are homing against a limit switch.
Here the motorRecord should not stop an ongoing homing, if MF_HOME_ON_LS
is set.

However, not stopping the motor from the motorRecord will confuse the
state machine in the .MIP field, and may result in an endless homing.

Solution:
Make the last change run-time configurable.
Use mr_stops_on_ls_activated() to determine if we need the "motorRecords stops"
or "motorRecord follows the DONE bit from the controller"