motorRecord.cc: Reduce spam printing of RA_DIRECTION
When the motor is standing still, but the encoder flickers,
the RA_DIRECTION bit in .MSTA may change often, spamming the
IOC log.
Reduce the number of prints:
If the motor "stands still", don't print.
Stand still is here defined as .MIP = DONE, and RA_MOVING moving is 0.
the old RA_MVING is 0, RA_DONE is 1 and old RA_DONE is 1