motorRecord.cc: Remove printing of RA_DIRECTION
Since the latest commit to reduce the priniting of
changed RA_DIRECTION sometimes misses a real update, remove the printing
of RA_DIRECTION all togother.
This bit influences only the .TDIR field, but does not have any impact
on the state machine otherwise. So printing it will not help anything.