asynMotorController.cpp: autopwerOff and no jog
When the motor should be jogged and autoPowerOff should be used,
it turns out that the motor is never powered off when the motor
never starts jogging.

This is a somewhat special case, the controller decides that it is
not a good time to jog, and does nothing.
Before doing nothing, the amplifier had been turned on,
and now it is never turned off.

Fix it by adding a call to setWasMovingFlag like this:
      status = pAxis->setClosedLoop(true);
 +    pAxis->setWasMovingFlag(1);

While there, replace "i" with "axisNo", that's what is is.