asynMotorAxis.cpp: Add positionWritten
When the motorPosition is yet unknown/undefined/never written
by the driver or poller, avoid callbacks into devMotorAsyn
Add a flag positionWritten:
When not set, callbacks are not (yet) done.
At some point in time, the position will be written by
setDoubleParam(pC_->motorPosition_) and then the callback is called.