Save-Restore: Add restore_needed
Make it possible to postpone the position-restore into motorRecord
processing, if needed.

When the position of the motor is unkown in init_record(), set
pmr->priv->saveRestore.restore_needed to 1.

The process() in motorRecord will pick it up once the controller
comes online, a callback is called and process() is run in motorRecord.cc