motorRecord: JOGF=1 and writing to e.g. LLM stopped
Fix an old bug with jogging:
When JOGF is set to 1, mip is set to MIP_JOG_REQ,
to signal that jogging needs to be commanded to the controller.
Once that is done, MIP_JOG_REQ bit is set to 0, MIP_JOGF to 1.

When the record is now processed because of e.g. writing to the LLM
field (or probably any other field), process() is executed.

The old code did stop the jogging because MIP_JOG_REQ==0.
Stopping the motor now is wrong.

This change keeps MIP_JOG_REQ to signal "need to request jogging to the
controller soon". Keep that logic.

Change the condition for a "STOP jogging":
When JOGF == 1 and MIP == MIP_JOGF, everything is fine, continue jogging,
don't stop the motor. (Same for JOGR).

When JOGF == 0 and MIP == MIP_JOGF the motor needs to be stopped.
Again, same for JOGR.

Cover even the case when JOGF == 1 and JOGR == 1.
Stop the motor, and once it is stopped, reset both JOGF and JOGR.