motorRecord.cc: More printing around commandedDval
When a new target position is given-
The motor is moving, and the .VAL field is changed,
this new value needs to be adressed after the motor has stopped.

There is currently a bug, when the NTM field is 1.
The motor starts at 0.0 and is commanded to 100.0
When the motor has passed 50.0, it is commanded to 25.0
Now it should be stopped (that is working) and after the stop
be moved to 25.0 (this is not working).
Start with improved printouts, and leave the fix for the next
commit.