Fix NTM=1 RTRY=1 bug: Motor stopped, but never moved
Setting .NTM to YES and didn't work:
The motor was stopped (when travelling into the "wrong" direction)
Once stopped, it was never comanded to move to the wanted position.

It did work with .RTRY == 0:
The maybeRetry() logic detected that the position is out ouf range,
no retry is enabled, resets the MIP bits.
Then doDVALchangedOrNOTdoneMoving() would pick up the new VAL/DVAL and
started the motor.

It did not work with .RTRY == 1:
The maybeRetry() logic detected that the position is out ouf range,
and .RTRY == 1 means "Too many retries".
Set .MISS and set SET_LAST_VAL_FROM_VAL, to prevent any further movements.

Solution:
Don't call maybeRetry() when the motor has stoppend and MIP is
MIP_STOP | MIP_MOVE