motorRecord/maybeRetry(): Do not retry if limit swicth is active
When maybeRetry() is called and the limit switch into the new direction
is active, don't start a move or retry.
This seems to happen only under weird conditions:
The motor is not homed, sitting on the low limit switch and reporting
a position > .HLM
From a position calculation the move should be possible.
But the limit switch says "no".

The record machinery now locks up in an endless retry loop, if .RTRY == 1

Add another helper variable, ls_active, print it and react on it.

As a second improvement, this should be detected in doRetryOrDone() and the
motor should not even be started. That's for another commit.