motorRecord.cc: Do not move further on limit switch
When a new position is commanded "towards" a limit switch that
is already activated, the record should not try to move the motor.

Example:
The driving range is 0..175 mm.
The controller is not homed and reports 201 mm.
The low limit switch is activated.
Trying to move the motor towards 0 is "legal" from a soft limit point
of view, but can not start because the low limit switch is already activated.

Avoid starting the motor:
Check in doMoveDialPosition() if the direction of move is towards
an activated limit switch, and if yes, don't start a movement.
Calculate pmr->cdir first, then check the appropriate limit switch.
If active, set the MIP to done.