asynMotorController.cpp: NULL pointer check for pasynUserController_
It may be that pasynUserController_ is left un-ininitailized,
because the driver doesn't use it.

Set it to NULL as an initialization and check for !NULL before using
it for printing.

Note: Using pasynUserController_ for generic printing is not 100% ideal:
pasynUserController_ is used by the octet interface, which most
controllers/drivers use (but not ecmc).

There are differnt solutions:
a) Remove the generic print in the poller for auto-power-on-off
b) Create a proper asynUser and use it
c) Use the NULL pointer check and don't print if pasynUserController_ != NULL

So for the moment go for c):
keep using pasynUserController_ for printing of e.g.
auto-power-on-off feature until we really need a better solution.