motorRecord.cc: If controller uses EGU, keep VELO when UREV changes
When MRES or UREV are changed, VELO, VBAS, BVEL, VMAX are adjusted
to the new resolution.
The fields S, SBAS. SBAK, SMAX are left as they are.

For controllers using EGU, do it the other way around:
Keep the values for VELO, VBAS, BVEL, VMAX and adjust
S, SBAS. SBAK, SMAX