motorRecord.cc: Do not move if VAL - RBV == RDBD
Improve a very corner case for the following scenario:
The motor is commanded to 170.0
It stops at exactly 170.1, which is RBV + RDBD

In this case a new movement should not be commanded.
The RDBD field says that the position should be "outside" the range.

Solution: Use ">=" instead of ">" when needed,
And "<" instead of "<=" where needed.