motorRecord: fall back to ACCS if JAR is 0.0
The acceleration for jogging is specified in the .JAR field.
When .JAR is 0.0, fall back to the value of the .ACCS field.

Setting .JAR to 0.0 is now allowed:
When jogging starts, use the current value of the .ACCS field.

Tryin to set .JAR < 0.0 will activate the old handling:
Set .JAR to .JVEL / 0.1