asynMotorController.cpp: make autoPowerOn() more rebust
Fix a race condition:
The poller() had just switched the amplifier off, calling
pAxis->setClosedLoop(false);
And at the same time a new movement should be started.
The race condition was here:
  getIntegerParam(axis, motorStatusPowerOn_, &motorStatusPowerOn);
Since motorStatusPowerOn still was true in the param library,
but the MCU had been instructed to power off the amplifier,
the now to be started movement fails.

Solution:
Do not use this shortcut that relies on the parameter library.
While there, simplify the code:
(autoPower == 2) shares code with (autoPower == 1)