motorRecord: loosen the close enough calculation
Loosen the calculation for the calculation if an axis
is inside RDBD a little bit:
Sometimes, due to roundings of floating point calculation,
the axis seems to be "outside".

The log shows only 6 digits of the floating point values, e.g.
"... rdbd=0.040000 diff=0.040000 ..."

In our setup the diff was very slightly out of range,
"2.046280e-14" beyond the border, and
if fabs(diff) <= rdbd
was false.

The solution is to loosen the calculation for the if statement and
to allow the diff to be 1% out of range