motorRecord.cc: Do not stop on ls_active
Revert an old change, that was introduced due to a misunderstanding:
When the motor hits a limit switch, the motorRecord should not stop
the motor.

The original story started here:
303a9208e368ef07839afff67bac69e429af7332
- Set "stop" field true if driver returns RA_PROBLEM true.

This was (partly) reverted in the master branch:
95c0a4caa22fd946f4b3064d72beb2837e5a845a

But the ess-master branch kept the feature that the .STOP was set
once a limit switch (towards the traveling direction) is activeted

In that sense: Follow the upstream branch here.