Introduce motorFlagsNotHomedProblem
Prepare the fix in motorRecord.cc:
When the problem bit is set, the motion needs to be
considered as "done".
The motor/master has always been like this, the ess-master not.
The main motivation had been that it should be possible to
set a "not-homed" motor(record) into alarm state major.
The only way to do this had been to set the PROBLEM bit,
(and as a result, PROBLEM == 1 does no longer mean "DONE")

Prepare everything to restore the old "PROBLEM is DONE" logic
in motorRecord while allowing to set the alarm state for
non-homed motors:

Convert motorFlagsNotHomedProblem into a flag:
Remove motorFlagsNotHomedProblem and add
motorFlagsNotHomedProblem

Change the alarm handling in motorRecord.cc to look at the new
flag and put the record into alarm state (when configured).