asynMotorAxis.cpp: Distinguish PowerOff vs PowerOff Auto
When the power of a motor is off, the message text shows "PowerOff"
However, it is not clear if the next movement will power it on
automatically because of the "PowerOnOff" feature.
Or if not, and a manual power on is needed.
Distinguish here: Either "PowerOff(Auto)" or "PowerOff" is shown.