RHLM, RLLM: Fix when MRES is negativ
Fix the situation when MRES < 0 and either the softlimits
are changed (and RHLM and RLLM are garbled) or when
MRES itself is negative and changed, so that DHLM and DLLM
needed to be updated from the constant RHLM/RLLM.
Based on the update of DHLM/DLLM HLM/LLM will be updated as well.

The current implementation around RHLM/RLLM did not consider
MRES < 0 at all. And even if this configuration is not often
used, it is still supported.
Beside that we want to avoid RHLM=-40 RLLM=50 when it should be
the other way around