motorRecord: Reset JOGR/JOGF when LLS/HLS is active
Reset the JOGR field, when jogging should be started,
but is inhibited by a limit switch.
Reset JOGR when LLS is active and JOGF when HLS is active.
This is a fix for a regression against motor/master,
where the JOGF/JOGR are reset as well, but in a different code path.