motorRecord.cc: doBackLash may hung up when on LS
When backlash should be done, it may be refused, because the motor
is already on a limit switch:

commit b47408f379ca87a11dff8a1b6832c867980c91d6
    motorRecord.cc: Do not move further on limit switch

However, as a result of the not-move, the motorRecord can get stuck
with DMOV == 0, when backlash is active and an activated limit switch
prevents a move.

Solution:
Add a check in doBackLash():
    if (moving_started)
    {
        /* Restore DMOV to false and UNMARK it so it is not posted. */
        pmr->dmov = FALSE;
        UNMARK(M_DMOV);
    }

While there, improve some of the printings to allow further debugging,
add missing MARK(M_MIP) and remove re-setting bits in the MIP which are not
set, to reduce the confusion when reading logs.
Things to be done in the next commits:
- Check that doMoveDialPosition() is not called to move from position X to X,
  the same position may be commanded when the first (backlash) move is rejected
  because of the activated limit switch, and the second movement is not needed.
- Check if the MARK(M_MIP) can be integrated into MIP_SET_VAL() and friends.