motorRecord: introduce too_small in doMoveDialPositionL()
When a movement with backlash should start, but the motor is already on
a limit switch, the second part of the movement may end up with something
like move from -20.0 to -20.0

Refuse these kind of movements in a "too_small" fashion.
The decision if to use SPDB or not goes like this, in opposite
to the other too_small functionality:
When SPDB is set and > 0.0, it should be used.
Only if SPDB is 0.0, MRES needs to be consulted.