motorRecord.cc: Do not move further when LS active
Use the existing logic for LVIO when sitting on a limit switch
and a new movement should be started.
Add another check, when the VAL (or DVAL, RVAl, TWF, TWR) fields had
been written to:
When a limit switch is active, and the direction is "further into the
limit switch", do not move.