asynMotorController: Do not move, when power on fails
Handle a special use case better:
The motor is normally stopped by an external hardware event, and powered off.
A new movement is requested, but is not allowed to happen.
The "auto power on" feature in the model 3 driver tries to power on
the motor, and the driver refuses to do so:
 pAxis->setClosedLoop(true) returns a failure.
This failure is not handled very well, first we try to loop until the
timeout has elapsed, and then the move is commanded anyway.

Solution:
Handle the return status of pAxis->setClosedLoop(true) and do not command
a move, if setClosedLoop(true) fails.