motorRecord.cc: blink DMOV when jogging does not start
This is a follow-up to
  commit 2420d29d2f35475778db61480558ae2519add06e
  Author: Torsten Bögershausen <torsten.bogershausen@esss.se>
  Date:   Thu Dec 21 14:56:30 2023 +0100
    motorRecord: Reset JOGR/JOGF when LLS/HLS is active

    Reset the JOGR field, when jogging should be started,
    but is inhibited by a limit switch.
    Reset JOGR when LLS is active and JOGF when HLS is active.
    This is a fix for a regression against motor/master,
    where the JOGF/JOGR are reset as well, but in a different code path

While testing the .DMOV field for blinking when a motion is commanded,
even if it is not started.
Originally written for the "drive fields" in motorRecord.cc:
VAL, DVAL, RVAL, RLV, TWF, TWR.
When reading the documentation about the DMOV field in motorRecord.html
we can find that even a write to JOGR or JOGF
should show the same behavior, DMOV goes 1-0-1.
This was missing, when a jog command was refused in the motorRecord
because the motor is alread on the limit switch, the soft limit is
reached or the motor is too close to it.
See "start_jog" in motorRecord.cc

In short: Add the missing "blink" even when jogging is never started.