docs/motorRecord.html: CNEN is implementation specific
The CNEN field has been developed with motion controllers which can
switch between open- and closed-loop.
In modern controllers this is hard-coded inside the motion controller,
or to be more exact, sometimes the configuration of ths controller.
In other words, the CNEN field has been used to enable or disable the
motor.
Make this different usage and functionality more clear.