motorRecord:Fix hang when RTRY == 0: SET_LAST_VALS_FROM_VALS
Do not hang when RTRY == 0 and BDST == 0
When the motor did not reach the target position,
a retry may be needed.
However, when RTRY == 0 and BDST == 0 the motorRecord
may hang by commanding a never ending loop of retries.
The root cause was a missing
pmr->priv->last.dval = pmr->dval;
in line 2146, where
pmr->priv->last.val = pmr->val;
had been written, last.val was updates twice and dval not at all.

Fix this and introduce a new macro SET_LAST_VALS_FROM_VALS.
This prevents that doDVALchangedOrNOTdoneMoving() was kicking in:

doDVALchangedOrNOTdoneMoving: begin ...
   commandedDval=96.000000 dval=96.000000 last.dval=48.000000 ...